JAMES-2132 moving MDN to a separated project
