+++ mdn/src/main/java/org/apache/james/mdn/DispositionActionMode.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.transport.mailets.jsieve.mdn;
+package org.apache.james.mdn;
 
 /**
  * Interface <code>DispositionActionMode</code> marks a type encapsulating
