+++ mdn/src/main/java/org/apache/james/mdn/MDNFactory.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.transport.mailets.jsieve.mdn;
+package org.apache.james.mdn;
 
 import javax.mail.MessagingException;
 import javax.mail.internet.MimeBodyPart;
