+++ mdn/src/main/java/org/apache/james/mdn/TypeFailed.java
@@ -17,13 +17,28 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.transport.mailets.jsieve.mdn;
+package org.apache.james.mdn;
+
 
 /**
- * Interface <code>DispositionSendingMode</code> marks a type encapsulating
- * disposition sending mode information as defined by RFC 2298.
- */
-public interface DispositionSendingMode
+ * Class <code>TypeFailed</code>
+ */    
+public class TypeFailed implements DispositionType
 {
 
+    /**
+     * Default Constructor
+     */
+    public TypeFailed()
+    {
+        super();
+    }
+    
+    /**
+     * @see java.lang.Object#toString()
+     */
+    public String toString()
+    {
+        return "failed";
+    }         
 }
